RestoreCommands!
Added new command to restore sound commands. This gets all the sound files in the resources folder that has the guild id on it, then recreates the command in the database based on those names. Bug fix for help command when db commands array is empty.

Also: Auto stash before merge of "dev" and "origin/dev"