Delete Recorded Commands
With the addition of recorded commands, there was really a nice way to delete those newly created commands. The only solution was to delete each individual command, resulting in the user typing out 3 separate delete commands. This new command fixes that. Pass in a 'root' command, and the slow and fast versions will be taken care of.